Changed id type

+++ src/main/kotlin/com/kriogenik/guzeeva/data/repositories/PersonRepository.kt
@@ -5,4 +5,4 @@ import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface PersonRepository: CrudRepository<Person, Long>
+interface PersonRepository: CrudRepository<Person, Int>
